requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.5
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.3
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1